Refactor reclaim-mannequin command
